+++ index.html
@@ -18,52 +18,8 @@
     </header>
  <div class="mainSection">
     <section class="leftSection">
-    <article class="taco"></article>
-   <!-- <div class="fishStuff">
-       <h3>Fish</h3>
-       <article class="fishList">
-           <section class="fishes">
-               <img src="https://cdn.pixabay.com/photo/2016/12/31/21/22/discus-fish-1943755__340.jpg" alt="" class="image">
-               <li>Species: Fishus Fishes</li>
-               <li>Name: Gilly</li>
-               <li>Length: 3 inches</li>
-               <li>Habitat: shallow water</li>
-               <li>Diet: plankton</li>
-           </section>
-           <section class="fishes">
-               <img src="https://cdn.pixabay.com/photo/2015/04/07/21/12/lionfish-711799__340.jpg" alt="" class="image">
-               <li>Species: Spinus Erectus</li>
-               <li>Name: Mufasa</li>
-               <li>Length: 12 inches</li>
-               <li>Habitat: warm ocean</li>
-               <li>Diet: fish</li>
-           </section>
-           <section class="fishes">
-               <img src="https://cdn.pixabay.com/photo/2014/06/23/06/00/longhorn-cowfish-375110__340.jpg" alt="" class="image">
-               <li>Species: Yellow Fish</li>
-               <li>Name: Big Burd</li>
-               <li>Length: 6 inches</li>
-               <li>Habitat: prefers water</li>
-               <li>Diet: crabs, coral</li>
-            </section>
-            <section class="fishes">
-                <img src="https://cdn.pixabay.com/photo/2019/08/09/15/30/tropical-4395272__340.jpg" alt="" class="image">
-                <li>Species: Tie Die</li>
-                <li>Name: Techno</li>
-                <li>Length: 6 inches</li>
-                <li>Habitat: warm reefs</li>
-                <li>Diet: coral</li>
-            </section>
-            <section class="fishes">
-                <img src="https://cdn.pixabay.com/photo/2016/08/31/13/56/fish-1633525__340.jpg" alt="" class="image">
-                <li>Species: Octopus</li>
-                <li>Name: Dan</li>
-                <li>Length: 13 inches</li>
-                <li>Habitat: warm shallow water</li>
-                <li>Diet: anything</li>
-            </section>
-        </article>
-    </div> -->
+     <article class="taco"></article>
+   
        <h3>Quotes</h3>
        <article class="fishQuotes">
            <section>
