+++ scripts/fish/fishDatabase.js
@@ -1,11 +1,11 @@
 
- export const database = {
+ export const fishDatabase = {
     fish: [
         {
             image: "https://cdn.pixabay.com/photo/2016/12/31/21/22/discus-fish-1943755__340.jpg",
             name: "Bart",
             species: "Biggie",
-            length: 3,
+            length: "3 inches",
             location:"shallow water",
             diet: "crustations",
         },
@@ -13,7 +13,7 @@
             image: "https://cdn.pixabay.com/photo/2016/12/31/21/22/discus-fish-1943755__340.jpg",
             name: "Gilly",
             species: "Fishus Fishes",
-            length: 3,
+            length: "3 inches",
             location:"shallow water",
             diet: "plankton"
         },
@@ -21,7 +21,7 @@
             image: "https://cdn.pixabay.com/photo/2015/04/07/21/12/lionfish-711799__340.jpg",
             name: "Mufasa",
             species: "Spinus Erectus",
-            length: 12,
+            length: "12 inches",
             location:"warm ocean",
             diet: "fish"
         },
@@ -29,7 +29,7 @@
             image: "https://cdn.pixabay.com/photo/2014/06/23/06/00/longhorn-cowfish-375110__340.jpg",
             name: "Big Burd",
             species: "Yellow Fish",
-            length: 6,
+            length: "6 inches",
             location:"mostly water",
             diet: "crabs, coral"
         },
@@ -37,7 +37,7 @@
             image: "https://cdn.pixabay.com/photo/2019/08/09/15/30/tropical-4395272__340.jpg",
             name: "Techno",
             species: "Tie Die",
-            length: 6,
+            length: "6 inches",
             location:"warm reefs",
             diet: "coral"
         },
@@ -45,9 +45,10 @@
             image: "https://cdn.pixabay.com/photo/2016/08/31/13/56/fish-1633525__340.jpg",
             name: "Dan",
             species: "Octopus",
-            length: 16,
+            length: "16 inches",
             location:"warm shallow waters",
             diet: "anything"
         }
     ]
+ 
 }
