+++ scripts/locations/locationsDatabase.js
@@ -1,0 +1,5 @@
+locations: [
+    {
+
+    }
+],
