+++ scripts/main.js
@@ -1,4 +1,4 @@
-    import {FishList} from "./FishList.js"
+    import {FishList} from "./fish/FishList.js"
 
     const parentHTMLElement = document.querySelector(".taco")
