+++ scripts/tips/tipsDatabase.js
@@ -1,0 +1,6 @@
+tips: [
+    {
+        
+    }
+    
+]
